Rename reporter files
This is a pre-step for the restructuring.
I will keep this commit in the hope that Git will recognize
files as renamed instead of deleted and newly created.
