+++ bandit/reporters/colored_base.h
@@ -1,9 +1,9 @@
-#ifndef BANDIT_COLORED_REPORTER_H
-#define BANDIT_COLORED_REPORTER_H
+#ifndef BANDIT_REPORTERS_COLORED_BASE_H
+#define BANDIT_REPORTERS_COLORED_BASE_H
 
 #include <ostream>
 #include <bandit/colorizers/colorizer.h>
-#include <bandit/reporters/progress_reporter.h>
+#include <bandit/reporters/progress_base.h>
 
 namespace bandit {
   namespace detail {
