+++ bandit/reporters/crash.h
@@ -1,9 +1,9 @@
-#ifndef BANDIT_CRASH_REPORTER_H
-#define BANDIT_CRASH_REPORTER_H
+#ifndef BANDIT_REPORTERS_CRASH_H
+#define BANDIT_REPORTERS_CRASH_H
 
 #include <iostream>
 #include <vector>
-#include <bandit/reporters/progress_reporter.h>
+#include <bandit/reporters/progress_base.h>
 
 namespace bandit {
   namespace detail {
