+++ specs/reporters/singleline.cpp
